pyproject.toml
[tool.poetry]
name = "mais"
version = "0.1.0"
description = ""
authors = ["RaphaelForesti <raphael.foresti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.5"}
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
taskipy = "^1.14.0"
ruff = "^0.7.4"
testcontainers = "^4.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_path = "."
addopts = ['-p no:warnings']

[tool-ruff]
line-length = 79
extend-exclude= ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run     = 'fastapi dev mais/app.py'

pre_test = 'task lint'
test     = 'pytest --cov=mais -vv --ignore=mais/settings.py'
post_test = 'coverage html'

lint    = 'ruff check . && ruff check . --diff'
format  = 'ruff check . --fix ruff format'